Fix Duna PDU fuel/waste capacity
The reactor in the legacy MKS PDU was the same as the 0.625m reactor part, but
with 25 times the fuel/waste storage capacity.  The NFE patch, however, changed
both parts to have the same capacity, 3 times that of the 0.625m reactor.  This
seems to have been a mistake; the MKS PDU's patched fuel/waste capacity
should've been 3 times its original capacity, again 25 times that of the 0.625m
reactor.

The Duna PDU's original fuel/waste capacity is 20% of the legacy MKS PDU, or 5
times that of the 0.625m reactor.  This change corrects its patched capacity to
match that ratio.  (Importantly, this allows it to hold more than about two days
of fuel.)