Fix inconsistent Duna PDU optimum temperature
The temperature targets defined in ModuleCoreHeat and FissionReactor ought to be
the same, so that the part avoids heating itself past the reactor's optimum
temperature.  I've changed the ModuleCoreHeat to match the FissionReactor,
rather than the other way around, because this way it matches the 0.625m reactor
that the Duna PDU is based on.