Disable Whirlijig's uranium extractor only when MKS is installed
Conversion from Ore to EnrichedUranium should be disabled when using MKS, which
has its own system for refining uranium.  But checking for an
UmbraSpaceIndustries folder is too broad, because that folder is also present
when using other USI mods, like USI-LS and Konstruction, that have nothing to do
with uranium.  Fortunately, MKS includes MM patches with :FOR[MKS] clauses,
which satisfy a :NEEDS[MKS] here.  This keeps the Whirlijig's uranium extractor
enabled when other USI mods are installed, and disables it only when using MKS
specifically.