Fix Duna_PDU and Tundra_PDU costs
The cost/power ratio of the NFE reactors, and patched USI reactors,
follows an approximate linear relationship.  I forgot to update the
costs when adapting the MKS_PDU2 patch for the Duna_PDU, though, and
that carried over to the Tundra_PDU as well.

Normally I'd follow the ratio between the unpatched costs of the two
parts, but that's not a useful guideline in this case since the Duna PDU
costs less but produces much more power.  Instead, I've brought both the
Duna and Tundra costs into line with the other reactors by interpolating
the (NFE-patched) costs of the USI_Nuke_125 and USI_Nuke_250 reactors.