Add USI resource sharing to fuel/waste containers
This adds two modules to the NFE nuclear fuel/waste containers that are
already present on their USI counterparts:

USI_ModuleResourceWarehouse enables logistics consumers to automatically
pull resources from the container on-demand.  Though it circumvents the
RadioactiveStorageContainer module's requirement for a skilled engineer
to perform a manual trasfer, it provides less player control (so there's
still a niche for manual transfers), and it's an important convenience
for MKS-style "kolonies" that consist of multiple disconnected vessels
and an implied logistics network -- especially since the MKS reactor
parts ("Duna" and "Tundra" PDUs) don't have much fuel capacity of their
own.

USI_ModuleRecycleBin allows the container to receive resources that
would otherwise be lost when a nearby part is disassembled ("recycled")
by an engineer.  Nothing seems to utilize this for nuclear resources
quite yet -- MKS allows parts to be disassembled only into MaterialKits,
not EnrichedUranium or DepletedFuel -- but the USI nuclear fuel/waste
containers have the module so there may be a purpose for it in the
future, and the NFE containers should be prepared for that.