Fix Whirlijig description in MKS patch
The Whirlijig's Ore->EnrichedUranium processor is disabled when using
MKS, because MKS has a different way of producing EnrichedUranium.  It
can be confusing if the part's description still mentions that converer,
though, so let's remove that part of the sentence with a regex.

(The second sentence still says "both processes", but that's OK since it
now refers to the EnrichedUranium and XenonGas outputs instead of the
Ore and DepletedFuel inputs.)