Add patch for USI 1.25m "short" reactor
This is an integration patch for a recently-added 1.25m USI nuclear
reactor that's approximately half the height of the other 1.25m USI
reactor.

Compared to its original definition, I've made the reactor significantly
more powerful in this patch because its original definition seems rather
unbalanced: unpatched, it has the same mass and cost as its full-size
counterpart, but produces only 17% as much power.  Its output is
one-third more than the (unpatched) 0.625m USI reactor, but that's less
than NFE's MX-EXP "KerboPower" 0.625m reactor.

So, instead of sticking to those ratios, I've made it half as powerful
as the (patched) full-size 1.25m USI reactor, as well as having half the
mass and cost.  This seems appropriate since it's half the size.  At
100EC/s output, it's a bit more powerful than the MX-EXP, which also
seems appropriate since this reactor is larger that one.