Added PatchManger Integration
Made the RTG and USI-reactor patches work with PatchManager - players will be able to activate/deactivate the patches using an in-game UI if they have PatchManager installed.

Behavior of the patches has not changed, and without PatchManager there should be no effect.  (Except for it being a bit harder to find the decaying RTG’s patch - note that it is not active by default.)